renamed "comp.js" to "composition.js"
